requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python
tabula-py
tabula-py[jpype]
camelot-py[base]
pydantic==1.10.13
rocketry